requirements.txt
djangorestframework==3.11.0
git+https://github.com/django/django.git@master#egg=django
django-money==1.1
Pygments==2.6.1
Markdown==3.2.1
coreapi==2.3.3
psycopg2-binary==2.8.4
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.0.1
PyYAML==5.3.1
django-sql-utils==0.5.0
